ai_agent/requirements.txt
google-adk
google-genai
fastapi
uvicorn
requests
python-dotenv